### 2017-12-12 #486 @jakevdp
@desilinguist – would you be willing to review this submission?
