### 2017-12-12 #486 @desilinguist
@jakevdp I'd be happy to review this but I am about to go on Christmas break so I can probably start when I'm back in early January?
