### 2017-12-12 #486 @jakevdp
@desilinguist – that should be fine. Thanks for being willing to review!
