### 2017-12-12 #501 @whedon
[REVIEW]: Category Encoders: a scikit-learn-contrib package of transformers for encoding categorical data
**Submitting author:** @wdm0006 (<a href="http://orcid.org/0000-0001-5188-3577">William McGinnis</a>)
**Repository:** <a href="https://github.com/scikit-learn-contrib/categorical-encoding" target ="_blank">https://github.com/scikit-learn-contrib/categorical-encoding</a>
**Version:** v1.2.5
