### 2017-12-12 #486 @jakevdp
@desilinguist – the review will take place over at #501. Thanks!
