### 2017-12-20 1a99d64
Initial commit for #40
