+++ category_encoders/tests/test_encoders.py
@@ -507,3 +507,56 @@ class TestEncoders(unittest.TestCase):
         enc.fit(X, y)
         self.verify_numeric(enc.transform(X_t))
         self.verify_numeric(enc.transform(X_t, y_t))
+
+    def test_target_encode_np(self):
+        """
+
+        :return:
+        """
+
+        X = self.create_array(n_rows=1000)
+        X_t = self.create_array(n_rows=100)
+        y = np.random.randn(X.shape[0])
+        y_t = np.random.randn(X_t.shape[0])
+
+        enc = encoders.TargetEncoder(verbose=1)
+        enc.fit(X, y)
+        self.verify_numeric(enc.transform(X_t))
+        self.verify_numeric(enc.transform(X_t, y_t))
+
+    def test_target_encode_out(self):
+        """
+
+        :return:
+        """
+
+        cols = ['C1', 'D', 'E', 'F']
+        X = self.create_dataset(n_rows=1000)
+        X_t = self.create_dataset(n_rows=100)
+        y = np.random.randn(X.shape[0])
+        y_t = np.random.randn(X_t.shape[0])
+
+        enc = encoders.TargetEncoder(verbose=1, cols=cols)
+        enc.fit(X, y)
+        self.verify_numeric(enc.transform(X_t))
+        self.verify_numeric(enc.transform(X_t, y_t))
+
+        enc = encoders.TargetEncoder(verbose=1)
+        enc.fit(X, y)
+        self.verify_numeric(enc.transform(X_t))
+        self.verify_numeric(enc.transform(X_t, y_t))
+
+        enc = encoders.TargetEncoder(verbose=1, drop_invariant=True)
+        enc.fit(X, y)
+        self.verify_numeric(enc.transform(X_t))
+        self.verify_numeric(enc.transform(X_t, y_t))
+
+        enc = encoders.TargetEncoder(verbose=1, return_df=False)
+        enc.fit(X, y)
+        self.assertTrue(isinstance(enc.transform(X_t), np.ndarray))
+        self.assertTrue(isinstance(enc.transform(X_t, y_t), np.ndarray))
+
+        enc = encoders.TargetEncoder(verbose=1, randomized=True, sigma=0.1)
+        enc.fit(X, y)
+        self.verify_numeric(enc.transform(X_t))
+        self.verify_numeric(enc.transform(X_t, y_t))
