+++ category_encoders/target_encoder.py
@@ -1,0 +1,237 @@
+"""Leave one out coding"""
+import numpy as np
+import pandas as pd
+from sklearn.base import BaseEstimator, TransformerMixin
+from category_encoders.utils import get_obj_cols, convert_input
+from sklearn.utils.random import check_random_state
+
+__author__ = 'chappers'
+
+
+class TargetEncoder(BaseEstimator, TransformerMixin):
+    def __init__(self, verbose=0, cols=None, drop_invariant=False, return_df=True, impute_missing=True,
+                 handle_unknown='impute', min_samples_leaf=1, smoothing=1):
+        """Target Encode for categorical features. Based on leave one out approach.
+
+    Parameters
+    ----------
+
+    verbose: int
+        integer indicating verbosity of output. 0 for none.
+    cols: list
+        a list of columns to encode, if None, all string columns will be encoded
+    drop_invariant: bool
+        boolean for whether or not to drop columns with 0 variance
+    return_df: bool
+        boolean for whether to return a pandas DataFrame from transform (otherwise it will be a numpy array)
+    impute_missing: bool
+        boolean for whether or not to apply the logic for handle_unknown, will be deprecated in the future.
+    handle_unknown: str
+        options are 'error', 'ignore' and 'impute', defaults to 'impute', which will impute the category -1. Warning: if
+        impute is used, an extra column will be added in if the transform matrix has unknown categories.  This can causes
+        unexpected changes in dimension in some cases.
+
+    min_samples_leaf : int
+        minimum samples to take category average into account
+    smoothing : int
+        smoothing effect to balance categorical average vs prior
+
+    Example
+    -------
+    >>>from category_encoders import *
+    >>>import pandas as pd
+    >>>from sklearn.datasets import load_boston
+    >>>bunch = load_boston()
+    >>>y = bunch.target
+    >>>X = pd.DataFrame(bunch.data, columns=bunch.feature_names)
+    >>>enc = TargetEncode(cols=['CHAS', 'RAD']).fit(X, y)
+    >>>numeric_dataset = enc.transform(X)
+    >>>print(numeric_dataset.info())
+
+    <class 'pandas.core.frame.DataFrame'>
+    RangeIndex: 506 entries, 0 to 505
+    Data columns (total 13 columns):
+    CRIM       506 non-null float64
+    ZN         506 non-null float64
+    INDUS      506 non-null float64
+    NOX        506 non-null float64
+    RM         506 non-null float64
+    AGE        506 non-null float64
+    DIS        506 non-null float64
+    TAX        506 non-null float64
+    PTRATIO    506 non-null float64
+    B          506 non-null float64
+    LSTAT      506 non-null float64
+    CHAS       506 non-null float64
+    RAD        506 non-null float64
+    dtypes: float64(13)
+    memory usage: 51.5 KB
+    None
+
+    References
+    ----------
+
+    .. [1] A Preprocessing Scheme for High-Cardinality Categorical Attributes in Classification and Prediction Problems. from
+    https://kaggle2.blob.core.windows.net/forum-message-attachments/225952/7441/high%20cardinality%20categoricals.pdf.
+        """
+        self.return_df = return_df
+        self.drop_invariant = drop_invariant
+        self.drop_cols = []
+        self.verbose = verbose
+        self.cols = cols
+        self.min_samples_leaf = min_samples_leaf
+        self.smoothing = smoothing
+        self._dim = None
+        self.mapping = None
+        self.impute_missing = impute_missing
+        self.handle_unknown = handle_unknown
+        self._mean = None
+        
+        
+    
+    def fit(self, X, y, **kwargs):
+        """Fit encoder according to X and y.
+        Parameters
+        ----------
+        X : array-like, shape = [n_samples, n_features]
+            Training vectors, where n_samples is the number of samples
[…]
