### 2017-12-20 2176d94
update init and tests
