+++ category_encoders/target_encoder.py
@@ -1,4 +1,4 @@
-"""Leave one out coding"""
+"""Target Encoder"""
 import numpy as np
 import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
