### 2017-12-20 #501 @wdm0006
Hello,

I had a general question about the nature of these papers. How should I handle extra authors? I was the original author and am sole maintainer so I put myself there, should I solicit other contributors to add themselves? Should contributors going forward add themselves? Etc.

Thank you, -Will
