### 2017-12-20 c365405
updating readme and docs
