+++ docs/source/targetencoder.rst
@@ -1,0 +1,5 @@
+Target Encoder
+==============
+
+.. autoclass:: category_encoders.target_encoder.TargetEncoder
+    :members:
