### 2018-01-08 #501 @arfon
> I had a general question about the nature of these papers. How should I handle extra authors? I was the original author and am sole maintainer so I put myself there, should I solicit other contributors to add themselves? Should contributors going forward add themselves? Etc.

We generally try not to police authorship here but in the past, package owners/maintainers have opened an issue on the repository (https://github.com/scikit-learn-contrib/categorical-encoding) to ask if other contributors would like to be authors on the paper.

@desilinguist - friendly reminder to get to this review when you get a chance 😄
