### 2018-01-08 #501 @wdm0006
Ok, cool, I'll open an issue then, that's a good idea.
