### 2018-01-08 89793d7
Added myself as contributor to paper.md
