### 2018-01-08 ff2e4a2
add inverse_transform
