+++ category_encoders/one_hot.py
@@ -167,7 +167,7 @@ class OneHotEncoder(BaseEstimator, TransformerMixin):
             raise ValueError('Unexpected input dimension %d, expected %d' % (X.shape[1], self._dim, ))
 
         if not self.cols:
-            return X
+            return X if self.return_df else X.values
 
         X = self.ordinal_encoder.transform(X)
 
@@ -182,8 +182,64 @@ class OneHotEncoder(BaseEstimator, TransformerMixin):
         else:
             return X.values
 
+    def inverse_transform(self, X_in):
+        """
+        Perform the inverse transformation to encoded data.
+
+        Parameters
+        ----------
+        X_in : array-like, shape = [n_samples, n_features]
+
+        Returns
+        -------
+        p: array, the same size of X_in
+
+        """
+        X = X_in.copy(deep=True)
+
+        # first check the type
+        X = convert_input(X)
+
+        if self._dim is None:
+            raise ValueError('Must train encoder before it can be used to inverse_transform data')
+
+        X = self.reverse_dummies(X, self.cols)
+
+        # then make sure that it is the right size
+        if X.shape[1] != self._dim:
+            if self.drop_invariant:
+                raise ValueError("Unexpected input dimension %d, the attribute drop_invariant should "
+                                 "set as False when transform data"%(X.shape[1],))
+            else:
+                raise ValueError('Unexpected input dimension %d, expected %d'% (X.shape[1], self._dim, ))
+
+        if not self.cols:
+            return X if self.return_df else X.values
+
+        if self.impute_missing and self.handle_unknown == 'impute':
+            for col in self.cols:
+                if any(X[col] == -1):
+                    raise ValueError("inverse_transform is not supported because transform impute "
+                                     "the unknown category -1 when encode %s"%(col,))
+
+        for switch in self.ordinal_encoder.mapping:
+            col_dict = {col_pair[1] : col_pair[0] for col_pair in switch.get('mapping')}
+            X[switch.get('col')] = X[switch.get('col')].apply(lambda x:col_dict.get(x))
+
+        return X if self.return_df else X.values
+
     def get_dummies(self, X_in, cols=None):
         """
+        Convert numerical variable into dummy variables
+
+        Parameters
+        ----------
+        X_in: DataFrame
+        cols: list-like, default None
+              Column names in the DataFrame to be encoded
+        Returns
+        -------
+        dummies : DataFrame
         """
 
         X = X_in.copy(deep=True)
@@ -210,5 +266,32 @@ class OneHotEncoder(BaseEstimator, TransformerMixin):
         # convert all of the bools into integers.
         for col in bin_cols:
             X[col] = X[col].astype(int)
+
+        return X
+
+    def reverse_dummies(self, X, cols):
+        """
+        Convert dummy variable into numerical variables
+
+        Parameters
+        ----------
+        X : DataFrame
+        cols: list-like
+              Column names in the DataFrame that be encoded
+
+        Returns
+        -------
+        numerical: DataFrame
+
+        """
+        out_cols = X.columns.values
+
+        for col in cols:
+            col_list = [col0 for col0 in out_cols if col0.startswith(col)]
+            value_array = np.array([int(col0.split('_')[1]) for col0 in col_list])
+            X[col] = np.dot(X[col_list].values, value_array.T)
+            out_cols = [col0 for col0 in out_cols if col0 not in col_list]
+
+        X = X.reindex(columns=out_cols + cols)
 
         return X

[…]
