### 2018-01-08 2d0c8a3
add inverse_transform
