### 2018-01-08 51dfbd4
Updated my author information
