+++ joss/paper.md
@@ -8,11 +8,21 @@ authors:
  - name: William D McGinnis
    orcid: 0000-0002-3009-9465
    affiliation: "1, 2"
+ - name: Chapman Siu
+   orcid: 0000-0002-2089-3796
+   affiliation: 3
+ - name: Andre S
+   orcid: 0000-0001-5104-0465
+   affiliation: 4
 affiliations:
  - name: Predikto, Inc.
    index: 1
  - name: Helton Tech, LLC
    index: 2
+ - name: Suncorp Group Ltd.
+   index: 3
+ - name: Jungle AI
+   index: 4
 date: 5 December 2017
 bibliography: paper.bib
 ---
@@ -25,7 +35,7 @@ provided by scikit-learn [@scikit-learn] itself, scipy[@scipy], pandas[@pandas],
 
 It includes a number of pre-existing encoders that are commonly used, notably Ordinal, Hashing and OneHot encoders [@idre][@carey][@hashing]. There are also some
 less frequently used encoders including Backward Difference, Helmert, Polynomial and Sum encoding [@idre][@carey]. Finally there are
-experimental encoders: LeaveOneOut, Binary[@andrethrill] and BaseN [@zhang][@onehot][@basen].
+experimental encoders: LeaveOneOut, Binary and BaseN [@zhang][@onehot][@basen].
 
 The goal of these sorts of transforms is to represent categorical data, which has no true order, as numeric values while
 balancing desires to keep the representation in as few dimensions as possible.  Category_encoders seeks to provide access
