### 2018-01-08 #501 @desilinguist
This is a very useful and well thought out piece of software, speaking as someone who has had to deal with categorical features a lot for various datasets. It's extremely helpful to have a single package that not only contains a large variety of methods to convert such categorical features into numerical ones but is also compatible with the leading Python machine learning package that is most widely used around the world.

I think the authors have addressed almost all aspects of the review except for a few minor shortcomings that can be easily addressed:

1. I think the statement of need can be improved. For experienced ML folks, it may be obvious why you would want to convert categorical features into numerical ones. However, it may be better to talk about that a bit more explicitly for more novice ML users and to also show a theoretical example, if possible.

2. Speaking of examples, I was hoping to find an easy to try actual example in the README and the front page of the documentation instead of just a pointer to the examples folder. In addition, the scripts in the examples folder actually don't show how to use these encoders with a Pandas dataframe which is a very nice feature to have and the authors should show that off more explicitly. I had to actually read the source code for some of the encoders to see examples of usage with dataframes which were embedded in the docstring (e.g., `BinaryEncoder`). Perhaps an actual Examples section in the documentation illustrating all different ways in which the package can be used (pandas input, pipeline support, pandas output, etc.) would serve much better?

Other than improving the documentation and examples, I have no other recommendations for this paper. I think it will prove quite useful to many folks.
