### 2018-01-08 #501 @wdm0006
Thanks for the feedback @desilinguist, I'll work on improving the readme and the statement of need.
