### 2018-01-08 #501 @jakevdp
Thanks for the prompt review, @desilinguist! @wdm0006 – let us know when the comments are addressed.
