### 2018-01-08 7cc1966
add info into paper.md
