### 2018-01-09 c41cce3
add inverse_transform
