### 2018-01-09 54bf7a3
add inverse_transform
