+++ joss/paper.md
@@ -38,7 +38,14 @@ Category_encoders is a scikit-learn-contrib module of transformers for encoding 
 module, category_encoders is fully compatible with the scikit-learn API [@scikit-learn-api]. It also uses heavily the tools
 provided by scikit-learn [@scikit-learn] itself, scipy[@scipy], pandas[@pandas], and statsmodels[@statsmodels].
 
-It includes a number of pre-existing encoders that are commonly used, notably Ordinal, Hashing and OneHot encoders [@idre][@carey][@hashing]. There are also some
+Categorical variables (wiki) are those that represent a fixed number of possible values, rather than a continuous number.  Each value assigns the measurement to one of those finite groups, or categories.  They differ from ordinal variables in that the distance from one category to another ought to be equal regardless of the number of categories, as opposed to ordinal variables which have some intrinsic ordering.  As an example:
+
+Ordinal: low, medium, high
+Categorical: Georgia, Alabama, South Carolina, … , New York
+
+The machine learning algorithms we will later use tend to want numbers, and not strings, as their inputs so we need some method of coding to convert them.
+
+Category_encoders includes a number of pre-existing encoders that are commonly used, notably Ordinal, Hashing and OneHot encoders [@idre][@carey][@hashing]. There are also some
 less frequently used encoders including Backward Difference, Helmert, Polynomial and Sum encoding [@idre][@carey]. Finally there are
 experimental encoders: LeaveOneOut, Binary and BaseN [@zhang][@onehot][@basen].
 
