### 2018-01-11 #501 @wdm0006
@desilinguist I've updated the README and added a little bit to the paper, please let me know if the changes are sufficient.
