### 2018-01-12 f30473f
add inverse_transform
