+++ category_encoders/binary.py
@@ -67,7 +67,9 @@ class BinaryEncoder(BaseEstimator, TransformerMixin):
     None
 
     """
-    def __init__(self, verbose=0, cols=None, drop_invariant=False, return_df=True, impute_missing=True, handle_unknown='impute'):
+
+    def __init__(self, verbose=0, cols=None, drop_invariant=False, return_df=True, impute_missing=True,
+                 handle_unknown='impute'):
         self.return_df = return_df
         self.drop_invariant = drop_invariant
         self.drop_cols = []
@@ -153,7 +155,7 @@ class BinaryEncoder(BaseEstimator, TransformerMixin):
 
         # then make sure that it is the right size
         if X.shape[1] != self._dim:
-            raise ValueError('Unexpected input dimension %d, expected %d' % (X.shape[1], self._dim, ))
+            raise ValueError('Unexpected input dimension %d, expected %d' % (X.shape[1], self._dim,))
 
         if not self.cols:
             return X if self.return_df else X.values
@@ -171,7 +173,7 @@ class BinaryEncoder(BaseEstimator, TransformerMixin):
         else:
             return X.values
 
-    def inverse_transform(self, X_in):
+    def inverse_transform(self, Xt):
         """
         Perform the inverse transformation to encoded data.
 
@@ -184,7 +186,7 @@ class BinaryEncoder(BaseEstimator, TransformerMixin):
         p: array, the same size of X_in
 
         """
-        X = X_in.copy(deep=True)
+        X = Xt.copy(deep=True)
 
         # first check the type
         X = convert_input(X)
@@ -198,9 +200,9 @@ class BinaryEncoder(BaseEstimator, TransformerMixin):
         if X.shape[1] != self._dim:
             if self.drop_invariant:
                 raise ValueError("Unexpected input dimension %d, the attribute drop_invariant should "
-                                 "set as False when transform data"%(X.shape[1],))
+                                 "set as False when transform data" % (X.shape[1],))
             else:
-                raise ValueError('Unexpected input dimension %d, expected %d'% (X.shape[1], self._dim, ))
+                raise ValueError('Unexpected input dimension %d, expected %d' % (X.shape[1], self._dim,))
 
         if not self.cols:
             return X if self.return_df else X.values
@@ -209,11 +211,11 @@ class BinaryEncoder(BaseEstimator, TransformerMixin):
             for col in self.cols:
                 if any(X[col] == -1):
                     raise ValueError("inverse_transform is not supported because transform impute "
-                                     "the unknown category -1 when encode %s"%(col,))
+                                     "the unknown category -1 when encode %s" % (col,))
 
         for switch in self.ordinal_encoder.mapping:
-            col_dict = {col_pair[1] : col_pair[0] for col_pair in switch.get('mapping')}
-            X[switch.get('col')] = X[switch.get('col')].apply(lambda x:col_dict.get(x))
+            col_dict = {col_pair[1]: col_pair[0] for col_pair in switch.get('mapping')}
+            X[switch.get('col')] = X[switch.get('col')].apply(lambda x: col_dict.get(x))
 
         return X if self.return_df else X.values
 
@@ -248,8 +250,8 @@ class BinaryEncoder(BaseEstimator, TransformerMixin):
             X[col] = X[col].map(lambda x: self.col_transform(x, digits))
 
             for dig in range(digits):
-                X[str(col) + '_%d' % (dig, )] = X[col].map(lambda r: int(r[dig]) if r is not None else None)
-                bin_cols.append(str(col) + '_%d' % (dig, ))
+                X[str(col) + '_%d' % (dig,)] = X[col].map(lambda r: int(r[dig]) if r is not None else None)
[…]

[…]
