### 2018-01-12 b84c0e9
add inverse_transform
