+++ category_encoders/basen.py
@@ -305,7 +305,7 @@ class BaseNEncoder(BaseEstimator, TransformerMixin):
                 value_array = np.array([int(col0.split('_')[1]) for col0 in col_list])
             else:
                 len0 = len(col_list)
-                value_array = np.array([base ** (len0 - 1 - i) for i in xrange(len0)])
+                value_array = np.array([base ** (len0 - 1 - i) for i in range(len0)])
 
             X[col] = np.dot(X[col_list].values, value_array.T)
             out_cols = [col0 for col0 in out_cols if col0 not in col_list]
