+++ category_encoders/tests/test_encoders.py
@@ -483,7 +483,7 @@ class TestEncoders(unittest.TestCase):
 
         enc = encoders.OneHotEncoder(verbose=1, return_df=False)
         enc.fit(X, None)
-        self.assertTrue(enc.transform(X_t[X_t['D'] != 'A']).shape[1] == 19)
+        self.assertTrue(enc.transform(X_t[X_t['D'] != 'A']).shape[1] == 18)
 
         enc = encoders.OneHotEncoder(verbose=1, return_df=True, impute_missing=True, handle_unknown='impute')
         enc.fit(X, None)
