### 2018-01-16 #501 @jakevdp
Thanks @desilinguist and @wdm0006!
