### 2018-01-16 #501 @jakevdp
@arfon, we're ready to accept this!
